+++ extension/Changelog.md
@@ -1,4 +1,35 @@
-# BCT 0.2.1 Changelog
-## Bugfixes:
+# BCT Changelog
+## 0.3.3
+### Features:
+- Added beta version
+
+## 0.3.2
+### Bugfixes:
+- Fixed a bug that the orgasm multiplicator only affected vibrators when the split arousal bar was active
+
+## 0.3.1
+### Bugfixes:
+- Fixed a bug that could trigger an orgasm while another one was already running, causing a chain of multiple orgasms
+- Fixed a bug that vibrations were not affected by the multiplicators
+
+## 0.3
+### Features:
+- Added a textbox showing the currently installed version in the settings menu
+- Added a button to access the changelog in the settings menu
+- Added a setting to position the arousal bar either below the orgasm bar or on the right of the character
+- Reworked settings menu
+- Show hints for settings by clicking on them
+- Show feedback when updating a tail
+
+## 0.2.1
+### Bugfixes:
 - Fixed a bug that caused the arousal bar to repeatedly blink (hopefully).
 - Fixed a bug that made the decay multipliers not work when BCE's alternate arousal was turned on.
+
+## 0.2
+### Features:
+- Added the settings menu
+- Added options to set multipliers for arousal/orgasm speed/decay
+- Added the option to turn the additional arousal bar on/off
+- Added the tail wagging module
+- Added the BC menu button hitbox fix